Fix postsubmit fails caused by https://github.com/istio/istio/pull/53318
