+++ releasenotes/notes/53318.yaml
@@ -36,9 +36,9 @@ upgradeNotes:
     with `helm upgrade istio-base` from this and all subsequent releases
     after running the below kubectl commands as a one-time migration:
 
-      - `kubectl label crds -l chart=istio "app.kubernetes.io/managed-by=Helm"`
-      - `kubectl annotate crds -l chart=istio "meta.helm.sh/release-name=istio-base"`
-      - `kubectl annotate crds -l chart=istio "meta.helm.sh/release-namespace=istio-system"` (replace with actual istio namespace)
+      - `kubectl label crds -l chart=istio -l app.kubernetes.io/part-of=istio "app.kubernetes.io/managed-by=Helm"`
+      - `kubectl annotate crds -l chart=istio -l app.kubernetes.io/part-of=istio "meta.helm.sh/release-name=istio-base"` (replace with actual `istio-base` Helm releasename)
+      - `kubectl annotate crds -l chart=istio -l app.kubernetes.io/part-of=istio "meta.helm.sh/release-namespace=istio-system"` (replace with actual istio namespace)
 
     If desired, the legacy non-Helm-upgradable CRDs and labels can be generated by setting `Values.base.enableCRDTemplates=false`
     during `helm install base`, but this option will be removed in a future release.
